refactor(package): add to utils
